V1.24
修复原本的多发子弹轮盘赌存在重复弹仓的bug，增加了不能禁言的单独语言。赛马的获胜积分奖励调整为获胜者获得所有下注积分。增加了吹牛游戏

+++ src/main/java/top/strelitzia/model/IntegralInfo.java
@@ -4,7 +4,7 @@ public class IntegralInfo {
     private long groupId;
     private Integer integral;
     private String name;
-    private Integer QQ;
+    private Long QQ;
     private Integer signCount;
 
     public long getGroupId() {
@@ -31,9 +31,9 @@ public class IntegralInfo {
         this.name = name;
     }
 
-    public Integer getQQ() { return QQ; }
+    public Long getQQ() { return QQ; }
 
-    public void setQQ(Integer QQ) { this.QQ = QQ; }
+    public void setQQ(Long QQ) { this.QQ = QQ; }
 
     public Integer getSignCount() { return signCount; }
